[418454] [Admin] Client API and UI for managing repositories in a server https://bugs.eclipse.org/bugs/show_bug.cgi?id=418454
Implementation of the server-side CDOAdmin, including a new Repository Configuration Manager implemented in a repository dedicated to server administration.  It provides local storage of XML configurations files for repositories added via the admin protocol and an app extension to find and start them on CDO server start-up.  It also provides server-wide authentication of the repository administration functions (create/delete repository) via its Administrator account.

Includes some refactorings:
  - pull up abstract authentication signal from CDO protocol
     into Net4j layer (includes pulling up NotAuthenticatedException
     and deprecating the API from CDO Common)
  - pull up authentication protocol from internal CDO protocol
  - add authentication challenge to CDO Admin protocol
Also distinguish between repositories that can and cannot be deleted because they are permanently configured in the server's XML configuration file.

On top of the revised CDOAdmin protocol are built some UI actions for
  - deletion of a repository, including confirmation dialog and error handling
  - creation of a repository (two-step wizard)
Both actions require authorization by providing the credentials of the admin repository's administrator user.

+++ plugins/org.eclipse.emf.cdo.server.admin/src/org/eclipse/emf/cdo/server/internal/admin/CDOAdminServer.java
@@ -7,6 +7,7 @@
  *
  * Contributors:
  *    Eike Stepper - initial API and implementation
+ *    Christian W. Damus (CEA LIST) - bug 418454
  */
 package org.eclipse.emf.cdo.server.internal.admin;
 
@@ -17,6 +18,7 @@ import org.eclipse.emf.cdo.server.IRepository;
 import org.eclipse.emf.cdo.server.internal.admin.bundle.OM;
 import org.eclipse.emf.cdo.server.internal.admin.protocol.CDOAdminServerProtocol;
 import org.eclipse.emf.cdo.server.spi.admin.CDOAdminHandler;
+import org.eclipse.emf.cdo.server.spi.admin.CDOAdminHandler2;
 import org.eclipse.emf.cdo.spi.common.admin.AbstractCDOAdmin;
 import org.eclipse.emf.cdo.spi.server.RepositoryFactory;
 
@@ -106,6 +108,12 @@ public class CDOAdminServer extends AbstractCDOAdmin
   protected boolean doCreateRepository(String name, String type, Map<String, Object> properties)
   {
     CDOAdminHandler handler = getAdminHandler(type);
+    if (handler instanceof CDOAdminHandler2)
+    {
+      // Must provide administrator credentials to create a repository
+      ((CDOAdminHandler2)handler).authenticateAdministrator();
+    }
+
     IRepository delegate = handler.createRepository(name, properties);
     CDOServerUtil.addRepository(container, delegate);
     return true;
@@ -122,9 +130,27 @@ public class CDOAdminServer extends AbstractCDOAdmin
     }
 
     IRepository delegate = repository.getDelegate();
+
+    // Can this handler delete the repository?
+    if (!canDelete(delegate, handler))
+    {
+      throw new IllegalStateException("The repository is permanently configured.");
+    }
+
+    if (handler instanceof CDOAdminHandler2)
+    {
+      // Must provide administrator credentials to delete a repository
+      ((CDOAdminHandler2)handler).authenticateAdministrator();
+    }
+
     LifecycleUtil.deactivate(delegate);
     handler.deleteRepository(delegate);
     return true;
+  }
+
+  protected boolean canDelete(IRepository repository, CDOAdminHandler handler)
+  {
+    return !(handler instanceof CDOAdminHandler2) || ((CDOAdminHandler2)handler).canDelete(repository);
   }
 
   @Override
